requirements.txt
allure-pytest~=2.13.5
pytest~=8.3.1
selenium~=4.23.1
webdriver-manager~=4.0.1
python-dotenv~=1.0.0
behave~=1.2.6
requests~=2.32.3
axe-selenium-python~=2.1.6